Split stats models from regressors creation and rename "get_*" functions into "create_*"
